requirements.txt
(Python 3.6.8)
numpy == 1.16.2
Keras == 2.2.4
Pillow == 6.0.0
PyQt5 == 5.12.2
opencv-python == 4.1.0.25
pytesseract == 0.2.6
scipy == 1.2.1
six == 1.12.0
tensorflow-gpu == 1.13.1
wheel == 0.33.1
Cython == 0.29.7
easygui == 0.98.1
h5py == 2.9.0

PyQt5Designer == 5.10.1
